change to new C++ addon interface
